Working version of Navier-Stokes solution with static and moving bodies

+++ src/plot_recipes.jl
@@ -19,7 +19,7 @@ const myblue = RGBA{Float64}(74/255,144/255,226/255,1)
   end
 end
 
-@recipe function plot(x::AbstractArray{S,1},y::AbstractArray{S,1},w::Fields.Nodes{T,NX,NY};trim::Int=0) where {S,T,NX,NY}
+@recipe function plot(x::AbstractArray{S,1},y::AbstractArray{S,1},w::Fields.Nodes{T,NX,NY};trim=0) where {S,T,NX,NY}
       grid --> :none
       ratio := 1
       linewidth --> 1
